// src/types/Repo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Repo = { id: string, owner_login: string, name: string, 
/**
 * RFC 3339. Set by the worker on insert.
 */
created_at: string, 
/**
 * RFC 3339. Updated by the worker on rename / metadata edit.
 */
updated_at: string, };